+++ src/routes/productRoutes.ts
@@ -5,19 +5,16 @@ import {
   createProduct,
   updateProduct,
   deleteProduct,
-} from "../controllers/productController.js";
+} from "../controllers/productController";
 import {
   favoriteProduct,
   unfavoriteProduct,
-} from "../controllers/likeController.js";
-import {
-  getComments,
-  createComment,
-} from "../controllers/commentController.js";
-import { asyncHandler } from "../middlewares/asyncHandler.js";
-import { authMiddleware } from "../middlewares/authMiddleware.js";
-import { checkOwnership } from "../middlewares/ownership.js";
-import { validateProduct } from "../middlewares/validateProduct.js";
+} from "../controllers/likeController";
+import { getComments, createComment } from "../controllers/commentController";
+import { asyncHandler } from "../middlewares/asyncHandler";
+import { authMiddleware } from "../middlewares/authMiddleware";
+import { checkOwnership } from "../middlewares/ownership";
+import { validateProduct } from "../middlewares/validateProduct";
 
 const router = express.Router();
